AB#106971 Follow new unit test patterns
